Split C++ code into multiple files
